+++ miden-vm/src/cli/prove.rs
@@ -146,14 +146,16 @@ impl ProveCmd {
         let proving_options = self.get_proof_options()?;
 
         // execute program and generate proof
-        let (stack_outputs, proof) = miden_prover::prove_sync(
-            &program,
-            stack_inputs,
-            advice_inputs,
-            &mut host,
-            proving_options,
-        )
-        .wrap_err("Failed to prove program")?;
+        let rt = tokio::runtime::Builder::new_current_thread().build().unwrap();
+        let (stack_outputs, proof) = rt
+            .block_on(miden_prover::prove(
+                &program,
+                stack_inputs,
+                advice_inputs,
+                &mut host,
+                proving_options,
+            ))
+            .wrap_err("Failed to prove program")?;
 
         println!("Program proved in {} ms", now.elapsed().as_millis());